Изменил способы и вид выводимых на стартовый дисплей данных

+++ SUP/Core/Inc/tasks/SensOut.h
@@ -19,8 +19,8 @@ void trueButtonEP();
 void trueButtonEM();
 
 uint8_t  battaryCharge(uint8_t battryType, uint8_t num_cell, uint16_t battary_voltage);
-int16_t expFiltrCharge(int16_t newVal, float k);
-int16_t expFiltrVbat(int16_t newVal, float k);
+int16_t expFiltrCharge(float newVal, float k);
+int16_t expFiltrVbat(float newVal, float k);
 void error_processing();
 void longButton();
 void StartSensOutTask(void *argument);
